src/processor/requirements.txt
boto3==1.35.2
python-dotenv==0.19.2
pymongo==4.8.0